Break out `pulumi about`
This commit breaks out the `about` command and its subcommands into a
new subpackage of `pkg/cmd/pulumi`. In doing so we are able to better
encapsulate the components of the command and more clearly delineate the
"APIs" of each command hierarchy.
